+++ app/settings/index/template.hbs
@@ -31,19 +31,6 @@
 
 <section class="Section">
 	<div class="Container">
-		<h2>Map</h2>
-		<p>Place your radio channel anywhere you feel like. If you do, it will appear on the {{link-to "map" "channels.map"}}.</p>
-		{{map-position-selection
-			submitTask=saveCoordinates
-			lat=(readonly model.coordinatesLatitude)
-			lng=(readonly model.coordinatesLongitude)
-		}}
-		<br>
-	</div>
-</section>
-
-<section class="Section">
-	<div class="Container">
 		<h2>Backup your radio</h2>
 		<p>You can export and save all data from this Radio4000 channel at any time. The following link will take you to a "json" text file that you can save and store. Preferably in a vault under the mountains. This file does not contain any private information.</p>
 		<p>
