+++ app/settings/map/template.hbs
@@ -1,0 +1,13 @@
+<section class="Section">
+	<div class="Container">
+		<h2>Map</h2>
+		<p>Place your radio channel anywhere you feel like. If you do, it will appear on the {{link-to "map" "channels.map"}}.</p>
+		{{map-position-selection
+			submitTask=saveCoordinates
+			lat=(readonly model.coordinatesLatitude)
+			lng=(readonly model.coordinatesLongitude)
+		}}
+		{{outlet}}
+	</div>
+</section>
+
