+++ app/settings/template.hbs
@@ -6,6 +6,7 @@
 		<div class="Container">
 			<nav class="Tabs Tabs--horizontal Tabs--animated">
 				{{#link-to "settings.index" class="Tabs-item"}}Channel{{/link-to}}
+				{{#link-to "settings.map" class="Tabs-item"}}Map{{/link-to}}
 				{{#link-to "settings.account" class="Tabs-item"}}Account{{/link-to}}
 			</nav>
 		</div>
